新增脚本配置

+++ continew-ai-front/src/main/java/top/charles7c/continew/admin/front/model/query/ModelScriptQuery.java
@@ -37,7 +37,7 @@ public class ModelScriptQuery implements Serializable {
      */
     @Schema(description = "模型名称")
     @Query(type = QueryType.EQ)
-    private Integer modelId;
+    private Long modelId;
 
     /**
      * 预设内容
